/subs endpoint removed
