+++ app.js
@@ -4,7 +4,6 @@ const path = require('path')
 
 const indexRouter = require('./routes/index')
 const moviesRouter = require('./routes/movies')
-const subsRouter = require('./routes/subtitles')
 
 const app = express()
 
@@ -17,7 +16,6 @@ app.use(cors());
 
 app.use('/', indexRouter)
 app.use('/movies', moviesRouter)
-app.use('/subs', subsRouter)
 
 app.listen(port, (err) => {
     console.log(`API listening on ${port}!`)
