+++ routes/movies.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const axios = require('axios')
+require('dotenv').config()
 
 router.get('/', async(req, res) => {
     moviesList = []
@@ -8,11 +9,13 @@ router.get('/', async(req, res) => {
         title: '',
         poster_url: '',
         link: '',
+        subtitleLink: '',
         torrents: [],
         source: 'YTS Torrents'
     }
     let response = await axios.get(`https://yts.mx/api/v2/list_movies.json?query_term=${req.query.q}`)
     console.log(response.data.data.movies)
+    var imdbId = response.data.data.movie.imdb_code
     if(response.data.data.movies == undefined) {
         res.sendStatus(404)
     } else {
@@ -27,17 +30,33 @@ router.get('/', async(req, res) => {
                     magnet: `magnet:?xt=urn:btih:${torrent.hash}&dn=${encodeURIComponent(movie.title)}`
                 })
             })
-            movieObject.title = movie.title
-            movieObject.link = movie.url
-            movieObject.poster_url = movie.medium_cover_image
-            moviesList.push(movieObject)
-            movieObject = {
-                title: '',
-                poster_url: '',
-                link: '',
-                torrents: [],
-                source: 'YTS Torrents'
-            }
+
+            const config = {
+                method: "get",
+                url: `https://api.opensubtitles.com/api/v1/subtitles?imdb_id=${imdbId}&languages=en`,
+                headers: {
+                    "Api-Key": `${process.env.OS_API_KEY}`
+                    }
+                }
+    
+                axios(config)
+                .then(resp=>{
+                    const url = resp.data.data[0].attributes.url
+                    movieObject.subtitleLink = url
+                    movieObject.title = movie.title
+                    movieObject.link = movie.url
+                    movieObject.poster_url = movie.medium_cover_image
+                    moviesList.push(movieObject)
+                    movieObject = {
+                        title: '',
+                        poster_url: '',
+                        link: '',
+                        subtitleLink: '',
+                        torrents: [],
+                        source: 'YTS Torrents'
+                    }
+                })
+                .catch(err=>console.log(err))
         })
         res.json(moviesList)
     }
